Separate components into files. Add routing
