+++ react-app/package.json
@@ -6,6 +6,7 @@
     "@material-ui/core": "^4.5.1",
     "react": "^16.10.2",
     "react-dom": "^16.10.2",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "3.2.0",
     "typeface-roboto": "0.0.75"
   },
