+++ react-app/public/index.html
@@ -39,5 +39,6 @@
       To begin the development, run `npm start` or `yarn start`.
       To create a production bundle, use `npm run build` or `yarn build`.
     -->
+    <script src="https://unpkg.com/react-bootstrap@next/dist/react-bootstrap.min.js"></script>
   </body>
 </html>
